updated the clean command
