+++ Makefile
@@ -40,11 +40,14 @@ watch:
 	@which watchman-make > /dev/null || ( echo 'install watchman' && exit 1 )
 	watchman-make -p 'src/**/*.js' 'src/*.js' 'test/**/*.js' 'test/*.js' -t ci
 
-clean:
+reset:
 	rm -rf flow-typed
 	rm -rf node_modules
 	rm -rf dist
 	rm -rf docs
 	make init
+
+clean:
+	rm -rf dist
 
 .PHONY: default watch ci init build clean docs type lint test
