+++ package.json
@@ -29,7 +29,7 @@
     "eslint-plugin-promise": "^3.5.0",
     "eslint-plugin-standard": "^3.0.1",
     "flow-bin": "0.49.1",
-    "flow-typed": "^2.0.0",
+    "flow-typed": "2.1.2",
     "jest": "^19.0.2",
     "npm-run-all": "^4.0.2",
     "pre-commit": "^1.2.2"
