+++ src/core/debug/index.js
@@ -1,0 +1,5 @@
+// @flow
+import { withFiles } from './-files'
+import { withRepl } from './-repl'
+
+export { withFiles, withRepl }
